Spaceports are a problem. Space ports allow ships building but there appears to have been legacy stuff here so that the "allowshipbuilding" tag was not always referenced. there were 2 spots that referenced tanslationindex 458 and not the allowshipbuilding tag. the bab5 mod has a space port that did not have the allowshipbuilding tag so at game start they could not build ships.
I had changed these 458 references to the allowshipsbuilding tag but that did not work for bab5 and i expect others as the mods spaceport did not have this tag.
so for now i made this do both.
look for the allow ship building tag and name==Space Port.

--HG--
branch : feature/Push here instead of develop