Attempt to fix scout looping. the theory is that the there can be situations where the ship can get into a situation where it it moves past it movement goals and gets into a loop where it will always miss everytime it comes around again. There maybe some other issue causing this. when looking at the effect for fleet ships the ship would appear to accelerate as it neared the target to past its FTL speed. This might somehow be an effect of the clock difference between player game time and ship time. Where player time is locked to real time and the ships are locked to an artificial time based on loops of the game.
I changed this in the scout code to slow down when approaching its target if its FTL speed is greater than the distance to the planet. The scout code is done in 3 speed groups. So it should never slow down to zero due to the effects at each speed group. hopefully it work but im not sure.

--HG--
branch : feature/Push here instead of develop