got the colony screen menu reset issue nailed. it was a threading issue but its been there for quite some time. it wasnt one of my threading things though its built in due to the GUI. So its tricky but it goes like this. this.shipyard=false; do a bunch of stuff. In this time something checks the value of shipyard like the colony screen checking to see if it can show a shiplist tab. then set the value of the shipyard to what it should be. foreach building in planet.buildinglist if there is a building that provides a shipyard then set shipyard to true.
the logic should fall through to false if there is no shipyard. it make sense in a normal logic flow but not in a multi threaded logic flow.

to fix this i use a temp value for shipyard while checking if it should be true.
tempship =false.
do all the stuff to check to see if it should be true. set it to true.
this.shipyard=tempship

i think i did an if then thing which is not needed i do not think but this way there the value of shipyard only changes when it changes. its still not "perfect". to be perfect the value should be locked. so it can not change while be checked but locking is a pain in the ass and in this case its not so important. worst case with this change is that there would be an undetectable delay when the shipyard was lost and the build menu removed the shiplist from the gui. should be impossible to make a mistake.

--HG--
branch : feature/Push here instead of develop