702 : https://bitbucket.org/CrunchyGremlin/sd-blackbox/downloads/BlackBox_RadicalElements_RC_0702.exe
changed ftl system reduction to be a divisor of the total normal ftl mod. So that + ftl speed from traits will be reduced by the ftl reduction sliders.
Made some small changes to the way AI defensive fleets are made and offensive ships are stored.
Made some small changes to the checking for ships in borders. that might need a little more work.

Also changed the way the in system FTL slider reduction works to fit in more with the docs changeable border node radius option. So that a ship will be in friendly space when it is in borders. meaning a ship can be in a system it owns but our side of its borders.Resulting in a different FTL speed depending on being in borders or not. this will really only be seen in mods that use that feature which you will have to wait for.
Note that this also is effected by allied and open boarder relations.

changed the FTL UI to not show double warp restriction icons when being effected by gravity wells and ftl reduction sliders.. it also changes color depending on the status of the FTL change.
It also shows FTL bonus over 100% so you can see when ftl bonus tech is taking effect.

--HG--
branch : RadicalElements