723 uploaded. https://bitbucket.org/CrunchyGremlin/sd-blackbox/downloads/BlackBox_RadicalElements_RC_0723.exe
this should prevent the game from crashing when it cant read the date in the save game header.
it should also not change the ship health on save load. (very old problem) im not quite sure on this one as its hard to test for me but i think i got it.
it also should not stack health on save load.
dont adjust projector radius on small maps.
when the game crashes try and switch to windowed mode before showing the exception tracker window. that may or may not work but fixes an old issue of being unable to see the exception on crash.
also added added a function to the F6 key to submit an issue without a crash. So if something isnt working right it should allow an issue to be filed from in game with some useful, to me, game state information.

--HG--
branch : release/july