upload build 206. more changes to ship research. ship design research will be more random and spread between hull types. The AI still focuses on ships that have most of the tech that it already has but it will do it a bit differently. chooses all ships that require the least tech to create. from that list it chooses a random hull weighted to the smallest. from the list it then chooses a random role wieghted to specialty roles like bombers, carriers, support and troopships. then chooses a random ship weighted to ships with the most tech needed to make.
also expanded the ship research to include platforms and stations. and freighters and scouts. this actually may cause a problem in some of the other AI areas. but so far i havent seen it.
the AI loves to build defensive platforms and such. i may need to tone that down a bit or at least control it some so more can be spent on ships.

look for the AI using a wider variety of ships.
having better ships using smaller hulls.

fixed a bug in supply shuttles when the hangar doesnt have a proper supply shuttle defined.

changed the ship ratios for build to be based on numbers of ships rather than a percentage of maintenance. eventually this will allow custom ship ratios for the ai.

--HG--
branch : develop