Add... if found idle colony ship make it colonize immediatly. This will make a planet colonize button order happen immediatly if a ship is available.
* Goal function for change to goal step. Allows ChangeToStep(stepName)

if colony ship detroyed or state is not in colonize mode cancel goal.
This allows the player to cancel the goal by making the colony ship do something else.

if colony goal target is not equal ship colonize target cancel goal. This allows the player to do a right click colonize on a new planet and have the current goal be removed from the empire.
there is a question here if the goal ship be removed or restarted but i think its intuitive to have it cancled. But it can easily to do either.